requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
# dotenv==0.0.5
filelock==3.9.0
flatbuffers==23.5.26
gast==0.4.0
gensim==4.3.2
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
LibRecommender==1.3.0
llvmlite==0.41.0
Markdown==3.4.4
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numba==0.58.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
protobuf==4.24.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scikit-surprise==1.1.3
scipy==1.11.2
six==1.16.0
smart-open==6.4.0
sqlparse==0.4.4
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# torch==2.0.1+cpu
# torchaudio==2.0.2+cpu
# torchvision==0.15.2+cpu
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.3.7
wrapt==1.15.0
